move all existing files to claimStatus subdir
