+++ build.sh
@@ -2,16 +2,16 @@
 
 echo "Using rollup to bundle & minify files..." 
 
-rollup -c rollup.config.mjs -i src/noRecordFound.js -o dist/noRecordFound.min.js
-rollup -c rollup.config.mjs -i src/appReceived.js -o dist/appReceived.min.js
-rollup -c rollup.config.mjs -i src/claimList.js -o dist/claimList.min.js
-rollup -c rollup.config.mjs -i src/claimDetail.js -o dist/claimDetail.min.js
-rollup -c rollup.config.mjs -i src/paymentDetail.js -o dist/paymentDetail.min.js
+rollup -c rollup.config.mjs -i src/claimStatus/noRecordFound.js -o dist/claimStatus/noRecordFound.min.js
+rollup -c rollup.config.mjs -i src/claimStatus/appReceived.js -o dist/claimStatus/appReceived.min.js
+rollup -c rollup.config.mjs -i src/claimStatus/claimList.js -o dist/claimStatus/claimList.min.js
+rollup -c rollup.config.mjs -i src/claimStatus/claimDetail.js -o dist/claimStatus/claimDetail.min.js
+rollup -c rollup.config.mjs -i src/claimStatus/paymentDetail.js -o dist/claimStatus/paymentDetail.min.js
 
 echo "Copying bundled scripts to Cypress test assets..."
-cp dist/noRecordFound.min.js cypress/fixtures/noRecordFound/noRecordFound.min.js
-cp dist/claimDetail.min.js cypress/fixtures/claimDetail/claimDetail.min.js
-cp dist/claimList.min.js cypress/fixtures/claimList/claimList.min.js
-cp dist/paymentDetail.min.js cypress/fixtures/paymentDetail/paymentDetail.min.js
+cp dist/claimStatus/noRecordFound.min.js cypress/fixtures/claimStatus/noRecordFound/noRecordFound.min.js
+cp dist/claimStatus/claimDetail.min.js cypress/fixtures/claimStatus/claimDetail/claimDetail.min.js
+cp dist/claimStatus/claimList.min.js cypress/fixtures/claimStatus/claimList/claimList.min.js
+cp dist/claimStatus/paymentDetail.min.js cypress/fixtures/claimStatus/paymentDetail/paymentDetail.min.js
 
 echo "Complete!"
