+++ cypress/e2e/claimStatus/noRecordFound.cy.js
@@ -3,7 +3,7 @@ describe("No Record Found page", () => {
     const now = new Date(2021, 3, 14); // month is 0-indexed
     cy.clock(now);
 
-    cy.visit("./cypress/fixtures/noRecordFound/noRecordFound.html");
+    cy.visit("./cypress/fixtures/claimStatus/noRecordFound/noRecordFound.html");
 
     cy.contains(
       "No record of a New Jersey State Plan Temporary Disability or State Plan Family Leave Insurance claim has been found under the social security number (666-00-0000)."
@@ -15,7 +15,7 @@ describe("No Record Found page", () => {
   });
 
   it("passes accessibility checks", () => {
-    cy.visit("./cypress/fixtures/noRecordFound/noRecordFound.html");
+    cy.visit("./cypress/fixtures/claimStatus/noRecordFound/noRecordFound.html");
     cy.checkBodyA11y();
   });
 });
