+++ cypress/fixtures/claimStatus/claimDetail/claimDetailEligible.html
