+++ cypress/fixtures/claimStatus/claimDetail/claimDetailEligibleEnded.html
