+++ cypress/fixtures/claimStatus/claimDetail/claimDetailEligibleProcessing.html
