+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleBlank.html
