+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleC10.html
