+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleC10M10.html
