+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleDDU.html
