+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleFLDU.html
