+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleInvalidWage2007.html
