+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligibleM10.html
