+++ cypress/fixtures/claimStatus/claimDetail/claimDetailIneligiblePrivatePlan.html
