+++ cypress/fixtures/claimStatus/claimDetail/claimDetailUndeterminedInProgress.html
