+++ cypress/fixtures/claimStatus/claimDetail/claimDetailUndeterminedInfoNeeded.html
