+++ cypress/fixtures/claimStatus/claimList/claimListNoRecent.html
