+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailLeaveEndedFLI.html
