+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailLeaveEndedTDI.html
