+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailMaxEntitlement.html
