+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailNoAdditional.html
