+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailP30ReceivedFLI.html
