+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailP30ReceivedTDI.html
