+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailP30Sent.html
