+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailPayCode996.html
