+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailRecovered.html
