+++ cypress/fixtures/claimStatus/paymentDetail/paymentDetailScheduled.html
