+++ cypress/fixtures/claimStatus/sharedAssets/lwd_logo.jpg
