+++ cypress/fixtures/claimStatus/sharedAssets/p30.png
