+++ src/claimStatus/appReceived.js
@@ -4,7 +4,7 @@ import {
   updateIcon,
   makeMobileFriendly,
   addFeedbackLink,
-} from "./modules/shared.mjs";
+} from "../modules/shared.mjs";
 
 if (document.readyState === "loading") {
   window.addEventListener("DOMContentLoaded", () => {
